+++ src/main/java/org/store/test/dto/ProductDto.java
@@ -2,9 +2,11 @@ package org.store.test.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductDto {
     private final long id;
     private final String name;
-    private final long cost;
+    private final BigDecimal cost;
 }
