+++ src/main/java/org/store/test/model/Product.java
@@ -2,13 +2,15 @@ package org.store.test.model;
 
 import lombok.EqualsAndHashCode;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode
 public class Product {
     private long id;
     private String name;
-    private long cost;
+    private BigDecimal cost;
 
-    public Product(long id, String name, long cost) {
+    public Product(long id, String name, BigDecimal cost) {
         this.id = id;
         this.name = name;
         this.cost = cost;
@@ -22,7 +24,7 @@ public class Product {
         return name;
     }
 
-    public long getCost() {
+    public BigDecimal getCost() {
         return cost;
     }
 }
